LPS-97276 Move impl to OSGi module.
